azure-local-dashboard/backend/requirements.txt
# Web Framework
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
gevent==24.11.1

# Authentication
PyJWT==2.10.1

# Remote Execution
pywinrm==0.5.0
paramiko==3.5.1
requests-ntlm==1.3.0

# AI Integration
anthropic==0.43.0

# Credential Encryption
cryptography==44.0.0

# Background Scheduling
APScheduler==3.11.0

# Data Validation
pydantic==2.10.5

# HTTP Client
requests==2.32.3
httpx==0.28.1

# Logging
python-json-logger==3.2.1

# Environment
python-dotenv==1.0.1

# Date/time
python-dateutil==2.9.0
